Extract error code into new module, refine messaging
